requirements.txt
flask>=3.0
mysql-connector-python>=8.0
python-dotenv>=1.0
gunicorn>=21.0
